updated 100-prime_factor.c

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -8,8 +8,8 @@
 
 int main(void)
 {
-	int num = 612852475143;
-	int count;
+	long int num = 612852475143;
+	long int count;
 
 	for (count = num - 1; count > 1; count--)
 	{
